backend/requirements.txt
Django>=3.2
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
drf-yasg
django-storages
boto3
Pillow
pytest
pytest-django
python-dotenv
